fixes folder structure
