--- template/vtex/src/components/Example/Example.stories.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { Example } from '.'
-
-export default {
-  title: 'Example',
-  component: Example,
-};
-
-export const Text = () => <Example text="Storybooks Text Example"/>
-
-export const Emoji = () => <Example text="😀 😎 👍 💯"/>
