--- template/vtex/src/components/Example/index.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react'
-
-interface Props {
-  text: string
-}
-
-export const Example = ({ text }: Props) => {
-  return <div>Example Component: {text}</div>
-}
